refactor(package): move around util classes
