+++ app/src/main/java/com/example/ezvault/viewmodel/FilterViewModel.java
@@ -7,7 +7,8 @@ import androidx.lifecycle.ViewModel;
 
 import com.example.ezvault.data.FilterRepository;
 import com.example.ezvault.model.Tag;
-import com.example.ezvault.utils.filter.MainItemFilter;
+import com.example.ezvault.model.utils.SortedItemListView;
+import com.example.ezvault.model.utils.filter.MainItemFilter;
 import com.example.ezvault.utils.UserManager;
 
 import java.text.SimpleDateFormat;
