+++ app/src/main/java/com/example/ezvault/viewmodel/ItemViewModel.java
@@ -11,10 +11,11 @@ import androidx.lifecycle.ViewModel;
 import com.example.ezvault.data.FilterRepository;
 import com.example.ezvault.model.Item;
 import com.example.ezvault.model.ItemList;
-import com.example.ezvault.view.utils.ItemListView;
-import com.example.ezvault.utils.filter.IItemFilter;
-import com.example.ezvault.utils.filter.ItemListFilter;
-import com.example.ezvault.utils.filter.MainItemFilter;
+import com.example.ezvault.model.utils.SortedItemListView;
+import com.example.ezvault.model.utils.ItemListView;
+import com.example.ezvault.model.utils.filter.IItemFilter;
+import com.example.ezvault.model.utils.filter.ItemListFilter;
+import com.example.ezvault.model.utils.filter.MainItemFilter;
 import com.example.ezvault.utils.UserManager;
 
 import java.util.ArrayList;
